--- a/.clang-tidy
+++ b/.clang-tidy
@@ -13,7 +13,6 @@ Checks: >
     -bugprone-forward-declaration-namespace,
     -bugprone-sizeof-expression,
     -bugprone-throw-keyword-missing,
-    -bugprone-unused-local-non-trivial-variable,
     -bugprone-incorrect-enable-if,
     -bugprone-switch-missing-default-case,
     -clang-analyzer-*,