--- a/.github/workflows/osrm-backend.yml
+++ b/.github/workflows/osrm-backend.yml
@@ -628,6 +628,13 @@ jobs:
       PR_NUMBER: ${{ github.event.pull_request.number }}
       GITHUB_REPOSITORY: ${{ github.repository }}
     steps: 
+      - name: Enable data.osm.pbf cache
+        uses: actions/cache@v4
+        with:
+          path: ~/data.osm.pbf
+          key: v1-data-osm-pbf
+          restore-keys: |
+            v1-data-osm-pbf
       - name: Enable compiler cache
         uses: actions/cache@v4
         with:
@@ -648,9 +655,15 @@ jobs:
           ref: ${{ github.head_ref }}
           path: pr
       - name: Install dependencies
-        run: |
-          python3 -m pip install "conan<2.0.0" "requests==2.31.0"
+        run: | 
+          python3 -m pip install "conan<2.0.0" "requests==2.31.0" "locust==2.28.0"
           sudo apt-get update -y && sudo apt-get install ccache
+      - name: Prepare data
+        run: |
+          if [ ! -f "~/data.osm.pbf" ]; then
+            wget http://download.geofabrik.de/europe/germany/berlin-latest.osm.pbf -O ~/data.osm.pbf
+          fi
+          gunzip -c ./pr/test/data/berlin_gps_traces.csv.gz > ~/gps_traces.csv
       - name: Prepare environment
         run: |
           echo "CCACHE_DIR=$HOME/.ccache" >> $GITHUB_ENV
@@ -690,6 +703,7 @@ jobs:
         run: |
           ccache -p
           ccache -s
+  
   ci-complete:
     runs-on: ubuntu-22.04
     needs: [build-test-publish, docker-image, windows-release-node, benchmarks]