--- a/.github/workflows/osrm-backend.yml
+++ b/.github/workflows/osrm-backend.yml
@@ -703,7 +703,7 @@ jobs:
         run: |
           ccache -p
           ccache -s
-  
+
   ci-complete:
     runs-on: ubuntu-22.04
     needs: [build-test-publish, docker-image, windows-release-node, benchmarks]