--- a/include/extractor/serialization.hpp
+++ b/include/extractor/serialization.hpp
@@ -215,7 +215,6 @@ inline void read(storage::tar::FileReader &reader,
                  const std::string &name,
                  detail::NameTableImpl<Ownership> &name_table)
 {
-    std::string buffer;
     util::serialization::read(reader, name, name_table.indexed_data);
 }
 } // namespace osrm::extractor::serialization