--- a/unit_tests/partitioner/bisection_graph_view.cpp
+++ b/unit_tests/partitioner/bisection_graph_view.cpp
@@ -16,12 +16,14 @@ using namespace osrm::util;
 
 BOOST_AUTO_TEST_SUITE(graph_view)
 
-static void shuffle(std::vector<EdgeWithSomeAdditionalData> &grid_edges)
+namespace {
+void shuffle(std::vector<EdgeWithSomeAdditionalData> &grid_edges)
 {
     std::random_device rd;
     std::mt19937 rng(rd());
     std::shuffle(grid_edges.begin(), grid_edges.end(), rng);
 }
+} // namespace
 
 BOOST_AUTO_TEST_CASE(separate_top_bottom)
 {