… third_party/sol2/include/sol/sol.hpp